+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Version 1.9 (not yet released)
 * Added possibility to create a specification collection for conditions.
 * Added condition to check multiple parameters.
 * Added action to find or create a campaign.
+* Added parameters: join date, start date, end date to the create membership actions.
 
 Version 1.8
 -----------